Default DataSHIELD configuration

Find the R packages that have DataSHIELD server configuration information in them and extract this information in a data frame of aggregation/assignment methods and a named list of R options. The DataSHIELD packages can be filtered by specifying explicitly the package names to be included or excluded. The package exclusion prevails over the inclusion.

defaultDSConfiguration(include = NULL, exclude = NULL)

Arguments

include
Character vector of package names to be explicitly included. If NULL, do not filter packages.

exclude
Character vector of package names to be explicitly excluded. If NULL, do not filter packages.

Examples

{
# detect DS packages
defaultDSConfiguration()
# exclude a DS package
defaultDSConfiguration(exclude="dsBase")
# include explicitely some DS packages
defaultDSConfiguration(include=c("dsBase", "dsOmics"))
}

#> $AggregateMethods
#>                        name                            value package version
#> 1               asFactorDS1              dsBase::asFactorDS1  dsBase   6.1.2
#> 2                  asListDS                 dsBase::asListDS  dsBase   6.1.2
#> 3               boxPlotGGDS              dsBase::boxPlotGGDS  dsBase   6.1.2
#> 4           checkNegValueDS          dsBase::checkNegValueDS  dsBase   6.1.2
#> 5                   classDS                  dsBase::classDS  dsBase   6.1.2
#> 6                 corTestDS                dsBase::corTestDS  dsBase   6.1.2
#> 7                     corDS                    dsBase::corDS  dsBase   6.1.2
#> 8                     covDS                    dsBase::covDS  dsBase   6.1.2
#> 9        dataFrameSubsetDS1       dsBase::dataFrameSubsetDS1  dsBase   6.1.2
#> 10            densityGridDS            dsBase::densityGridDS  dsBase   6.1.2
#> 11                    dimDS                    dsBase::dimDS  dsBase   6.1.2
#> 12                   glmDS1                   dsBase::glmDS1  dsBase   6.1.2
#> 13                   glmDS2                   dsBase::glmDS2  dsBase   6.1.2
#> 14             glmerSLMADS2             dsBase::glmerSLMADS2  dsBase   6.1.2
#> 15          glmPredictDS.ag          dsBase::glmPredictDS.ag  dsBase   6.1.2
#> 16               glmSLMADS1               dsBase::glmSLMADS1  dsBase   6.1.2
#> 17               glmSLMADS2               dsBase::glmSLMADS2  dsBase   6.1.2
#> 18          glmSummaryDS.ag          dsBase::glmSummaryDS.ag  dsBase   6.1.2
#> 19            heatmapPlotDS            dsBase::heatmapPlotDS  dsBase   6.1.2
#> 20             histogramDS1             dsBase::histogramDS1  dsBase   6.1.2
#> 21             histogramDS2             dsBase::histogramDS2  dsBase   6.1.2
#> 22                     lsDS                     dsBase::lsDS  dsBase   6.1.2
#> 23                   isNaDS                   dsBase::isNaDS  dsBase   6.1.2
#> 24                isValidDS                dsBase::isValidDS  dsBase   6.1.2
#> 25              kurtosisDS1              dsBase::kurtosisDS1  dsBase   6.1.2
#> 26              kurtosisDS2              dsBase::kurtosisDS2  dsBase   6.1.2
#> 27                 lengthDS                 dsBase::lengthDS  dsBase   6.1.2
#> 28                 levelsDS                 dsBase::levelsDS  dsBase   6.1.2
#> 29                 lexisDS1                 dsBase::lexisDS1  dsBase   6.1.2
#> 30 listDisclosureSettingsDS dsBase::listDisclosureSettingsDS  dsBase   6.1.2
#> 31              lmerSLMADS2              dsBase::lmerSLMADS2  dsBase   6.1.2
#> 32             matrixDetDS1             dsBase::matrixDetDS1  dsBase   6.1.2
#> 33                   meanDS                   dsBase::meanDS  dsBase   6.1.2
#> 34               meanSdGpDS               dsBase::meanSdGpDS  dsBase   6.1.2
#> 35                messageDS                dsBase::messageDS  dsBase   6.1.2
#> 36                  namesDS                  dsBase::namesDS  dsBase   6.1.2
#> 37                  numNaDS                  dsBase::numNaDS  dsBase   6.1.2
#> 38           quantileMeanDS           dsBase::quantileMeanDS  dsBase   6.1.2
#> 39                  rangeDS                  dsBase::rangeDS  dsBase   6.1.2
#> 40                     rmDS                     dsBase::rmDS  dsBase   6.1.2
#> 41            scatterPlotDS            dsBase::scatterPlotDS  dsBase   6.1.2
#> 42              scoreVectDS              dsBase::scoreVectDS  dsBase   6.1.2
#> 43                setSeedDS                dsBase::setSeedDS  dsBase   6.1.2
#> 44              skewnessDS1              dsBase::skewnessDS1  dsBase   6.1.2
#> 45              skewnessDS2              dsBase::skewnessDS2  dsBase   6.1.2
#> 46                table1DDS                dsBase::table1DDS  dsBase   6.1.2
#> 47                table2DDS                dsBase::table2DDS  dsBase   6.1.2
#> 48                  tableDS                  dsBase::tableDS  dsBase   6.1.2
#> 49                 tableDS2                 dsBase::tableDS2  dsBase   6.1.2
#> 50                 tapplyDS                 dsBase::tapplyDS  dsBase   6.1.2
#> 51          testObjExistsDS          dsBase::testObjExistsDS  dsBase   6.1.2
#> 52                tTestFDS2                dsBase::tTestFDS2  dsBase   6.1.2
#> 53                    varDS                    dsBase::varDS  dsBase   6.1.2
#> 54               colnamesDS               dsBase::colnamesDS  dsBase   6.1.2
#> 55                   exists                     base::exists  dsBase   6.1.2
#> 56             is.character               base::is.character  dsBase   6.1.2
#> 57                is.factor                  base::is.factor  dsBase   6.1.2
#> 58                  is.list                    base::is.list  dsBase   6.1.2
#> 59                  is.null                    base::is.null  dsBase   6.1.2
#> 60               is.numeric                 base::is.numeric  dsBase   6.1.2
#> 61                     NROW                       base::NROW  dsBase   6.1.2
#> 62                   t.test                    stats::t.test  dsBase   6.1.2
#>         type    class
#> 1  aggregate function
#> 2  aggregate function
#> 3  aggregate function
#> 4  aggregate function
#> 5  aggregate function
#> 6  aggregate function
#> 7  aggregate function
#> 8  aggregate function
#> 9  aggregate function
#> 10 aggregate function
#> 11 aggregate function
#> 12 aggregate function
#> 13 aggregate function
#> 14 aggregate function
#> 15 aggregate function
#> 16 aggregate function
#> 17 aggregate function
#> 18 aggregate function
#> 19 aggregate function
#> 20 aggregate function
#> 21 aggregate function
#> 22 aggregate function
#> 23 aggregate function
#> 24 aggregate function
#> 25 aggregate function
#> 26 aggregate function
#> 27 aggregate function
#> 28 aggregate function
#> 29 aggregate function
#> 30 aggregate function
#> 31 aggregate function
#> 32 aggregate function
#> 33 aggregate function
#> 34 aggregate function
#> 35 aggregate function
#> 36 aggregate function
#> 37 aggregate function
#> 38 aggregate function
#> 39 aggregate function
#> 40 aggregate function
#> 41 aggregate function
#> 42 aggregate function
#> 43 aggregate function
#> 44 aggregate function
#> 45 aggregate function
#> 46 aggregate function
#> 47 aggregate function
#> 48 aggregate function
#> 49 aggregate function
#> 50 aggregate function
#> 51 aggregate function
#> 52 aggregate function
#> 53 aggregate function
#> 54 aggregate function
#> 55 aggregate function
#> 56 aggregate function
#> 57 aggregate function
#> 58 aggregate function
#> 59 aggregate function
#> 60 aggregate function
#> 61 aggregate function
#> 62 aggregate function
#> 
#> $AssignMethods
#>                                  name
#> 1                               absDS
#> 2                       asCharacterDS
#> 3                      asDataMatrixDS
#> 4                          asFactorDS
#> 5                         asFactorDS2
#> 6                         asIntegerDS
#> 7                            asListDS
#> 8                         asLogicalDS
#> 9                          asMatrixDS
#> 10                        asNumericDS
#> 11                            BooleDS
#> 12         boxPlotGG_data_TreatmentDS
#> 13 boxPlotGG_data_Treatment_numericDS
#> 14                            cbindDS
#> 15                                cDS
#> 16                   changeRefGroupDS
#> 17                    completeCasesDS
#> 18                        dataFrameDS
#> 19                    dataFrameFillDS
#> 20                    dataFrameSortDS
#> 21                 dataFrameSubsetDS2
#> 22                 glmerSLMADS.assign
#> 23                    glmPredictDS.as
#> 24                   glmSLMADS.assign
#> 25                    glmSummaryDS.as
#> 26                          getWGSRDS
#> 27                             listDS
#> 28                           lexisDS2
#> 29                           lexisDS3
#> 30                  lmerSLMADS.assign
#> 31                       matrixDetDS2
#> 32                       matrixDiagDS
#> 33                   matrixDimnamesDS
#> 34                           matrixDS
#> 35                     matrixInvertDS
#> 36                       matrixMultDS
#> 37                  matrixTransposeDS
#> 38                            mergeDS
#> 39                            rbindDS
#> 40                           rBinomDS
#> 41                     recodeLevelsDS
#> 42                     recodeValuesDS
#> 43                              repDS
#> 44                        replaceNaDS
#> 45                          reShapeDS
#> 46                            rNormDS
#> 47                       rowColCalcDS
#> 48                            rPoisDS
#> 49                            rUnifDS
#> 50                           sampleDS
#> 51                             seedDS
#> 52                              seqDS
#> 53                             sqrtDS
#> 54                    subsetByClassDS
#> 55                           subsetDS
#> 56                     tableDS.assign
#> 57                    tapplyDS.assign
#> 58                           unListDS
#> 59                       as.character
#> 60                            as.null
#> 61                         as.numeric
#> 62                             attach
#> 63                                  c
#> 64                              cbind
#> 65                     complete.cases
#> 66                                exp
#> 67                               list
#> 68                                log
#> 69                                rep
#> 70                                sum
#> 71                             unlist
#>                                         value package version   type    class
#> 1                               dsBase::absDS  dsBase   6.1.2 assign function
#> 2                       dsBase::asCharacterDS  dsBase   6.1.2 assign function
#> 3                      dsBase::asDataMatrixDS  dsBase   6.1.2 assign function
#> 4                          dsBase::asFactorDS  dsBase   6.1.2 assign function
#> 5                         dsBase::asFactorDS2  dsBase   6.1.2 assign function
#> 6                         dsBase::asIntegerDS  dsBase   6.1.2 assign function
#> 7                            dsBase::asListDS  dsBase   6.1.2 assign function
#> 8                         dsBase::asLogicalDS  dsBase   6.1.2 assign function
#> 9                          dsBase::asMatrixDS  dsBase   6.1.2 assign function
#> 10                        dsBase::asNumericDS  dsBase   6.1.2 assign function
#> 11                            dsBase::BooleDS  dsBase   6.1.2 assign function
#> 12         dsBase::boxPlotGG_data_TreatmentDS  dsBase   6.1.2 assign function
#> 13 dsBase::boxPlotGG_data_Treatment_numericDS  dsBase   6.1.2 assign function
#> 14                            dsBase::cbindDS  dsBase   6.1.2 assign function
#> 15                                dsBase::cDS  dsBase   6.1.2 assign function
#> 16                   dsBase::changeRefGroupDS  dsBase   6.1.2 assign function
#> 17                    dsBase::completeCasesDS  dsBase   6.1.2 assign function
#> 18                        dsBase::dataFrameDS  dsBase   6.1.2 assign function
#> 19                    dsBase::dataFrameFillDS  dsBase   6.1.2 assign function
#> 20                    dsBase::dataFrameSortDS  dsBase   6.1.2 assign function
#> 21                 dsBase::dataFrameSubsetDS2  dsBase   6.1.2 assign function
#> 22                 dsBase::glmerSLMADS.assign  dsBase   6.1.2 assign function
#> 23                    dsBase::glmPredictDS.as  dsBase   6.1.2 assign function
#> 24                   dsBase::glmSLMADS.assign  dsBase   6.1.2 assign function
#> 25                    dsBase::glmSummaryDS.as  dsBase   6.1.2 assign function
#> 26                          dsBase::getWGSRDS  dsBase   6.1.2 assign function
#> 27                             dsBase::listDS  dsBase   6.1.2 assign function
#> 28                           dsBase::lexisDS2  dsBase   6.1.2 assign function
#> 29                           dsBase::lexisDS3  dsBase   6.1.2 assign function
#> 30                  dsBase::lmerSLMADS.assign  dsBase   6.1.2 assign function
#> 31                       dsBase::matrixDetDS2  dsBase   6.1.2 assign function
#> 32                       dsBase::matrixDiagDS  dsBase   6.1.2 assign function
#> 33                   dsBase::matrixDimnamesDS  dsBase   6.1.2 assign function
#> 34                           dsBase::matrixDS  dsBase   6.1.2 assign function
#> 35                     dsBase::matrixInvertDS  dsBase   6.1.2 assign function
#> 36                       dsBase::matrixMultDS  dsBase   6.1.2 assign function
#> 37                  dsBase::matrixTransposeDS  dsBase   6.1.2 assign function
#> 38                            dsBase::mergeDS  dsBase   6.1.2 assign function
#> 39                            dsBase::rbindDS  dsBase   6.1.2 assign function
#> 40                           dsBase::rBinomDS  dsBase   6.1.2 assign function
#> 41                     dsBase::recodeLevelsDS  dsBase   6.1.2 assign function
#> 42                     dsBase::recodeValuesDS  dsBase   6.1.2 assign function
#> 43                              dsBase::repDS  dsBase   6.1.2 assign function
#> 44                        dsBase::replaceNaDS  dsBase   6.1.2 assign function
#> 45                          dsBase::reShapeDS  dsBase   6.1.2 assign function
#> 46                            dsBase::rNormDS  dsBase   6.1.2 assign function
#> 47                       dsBase::rowColCalcDS  dsBase   6.1.2 assign function
#> 48                            dsBase::rPoisDS  dsBase   6.1.2 assign function
#> 49                            dsBase::rUnifDS  dsBase   6.1.2 assign function
#> 50                           dsBase::sampleDS  dsBase   6.1.2 assign function
#> 51                             dsBase::seedDS  dsBase   6.1.2 assign function
#> 52                              dsBase::seqDS  dsBase   6.1.2 assign function
#> 53                             dsBase::sqrtDS  dsBase   6.1.2 assign function
#> 54                    dsBase::subsetByClassDS  dsBase   6.1.2 assign function
#> 55                           dsBase::subsetDS  dsBase   6.1.2 assign function
#> 56                     dsBase::tableDS.assign  dsBase   6.1.2 assign function
#> 57                    dsBase::tapplyDS.assign  dsBase   6.1.2 assign function
#> 58                           dsBase::unListDS  dsBase   6.1.2 assign function
#> 59                         base::as.character  dsBase   6.1.2 assign function
#> 60                              base::as.null  dsBase   6.1.2 assign function
#> 61                           base::as.numeric  dsBase   6.1.2 assign function
#> 62                               base::attach  dsBase   6.1.2 assign function
#> 63                                    base::c  dsBase   6.1.2 assign function
#> 64                                base::cbind  dsBase   6.1.2 assign function
#> 65                      stats::complete.cases  dsBase   6.1.2 assign function
#> 66                                  base::exp  dsBase   6.1.2 assign function
#> 67                                 base::list  dsBase   6.1.2 assign function
#> 68                                  base::log  dsBase   6.1.2 assign function
#> 69                                  base::rep  dsBase   6.1.2 assign function
#> 70                                  base::sum  dsBase   6.1.2 assign function
#> 71                               base::unlist  dsBase   6.1.2 assign function
#> 
#> $Options
#> $Options$datashield.privacyLevel
#> [1] "5"
#> 
#> $Options$default.nfilter.glm
#> [1] "0.33"
#> 
#> $Options$default.nfilter.kNN
#> [1] "3"
#> 
#> $Options$default.nfilter.string
#> [1] "80"
#> 
#> $Options$default.nfilter.subset
#> [1] "3"
#> 
#> $Options$default.nfilter.stringShort
#> [1] "20"
#> 
#> $Options$default.nfilter.tab
#> [1] "3"
#> 
#> $Options$default.nfilter.noise
#> [1] "0.25"
#> 
#> $Options$default.nfilter.levels
#> [1] "0.33"
#> 
#>