Setup an environment based on named datasets and logindata

Load the provided datasets and the corresponding logindata object, instanciate a new DSLiteServer hosting these datasets, verifies that the required DataSHIELD server-side packages are installed. All the data structures are loaded by data which supports various formats (see data() documentation).

setupDSLiteServer(
  packages = c(),
  datasets,
  logindata,
  pkgs = NULL,
  dslite.server = NULL,
  env = parent.frame()
)

Arguments

packages

DataSHIELD server-side packages which local installation must be verified so that the DSLiteServer can auto-configure itself and can execute the DataSHIELD operations. Default is none.

datasets

Names of the datasets to be loaded using data.

logindata

Name of the login data object to be loaded using data.

pkgs

The package(s) to look in for datasets, default is all, then the 'data' subdirectory (if present) of the current working directory (same behavior as 'package' argument in data).

dslite.server

Symbol name to which the DSLiteServer should be assigned to. If not provided, the symbol name will be the first not null one specified in the 'url' column of the loaded login data.

env

The environment where DataSHIELD objects should be looked for: the DSLiteServer and the DSIConnection objects. Default is the Global environment.

Value

The login data for the datashield.login function.

See also

Examples

{
logindata <- setupDSLiteServer(
                 datasets = c("CNSIM1", "CNSIM2", "CNSIM3"),
                 logindata = "logindata.dslite.cnsim", pkgs = "DSLite",
                 dslite.server = "dslite.server")
conns <- datashield.login(logindata, assign=TRUE)
# do DataSHIELD analysis
datashield.logout(conns)
}
#> 
#> Logging into the collaborating servers
#> 
#>   No variables have been specified. 
#>   All the variables in the table 
#>   (the whole dataset) will be assigned to R!
#> 
#> Assigning table data...