Setup a test environment based on the SURVIVAL (EXPAND_WITH_MISSING) simulated datasets

Load the SURVIVAL (EXPAND_WITH_MISSING) datasets, the corresponding login data object, instanciate a new DSLiteServer hosting these datasets and verify that the required DataSHIELD server-side packages are installed.

setupSURVIVALTest(packages = c(), env = parent.frame())

Arguments

packages

DataSHIELD server-side packages which local installation must be verified so that the DSLiteServer can auto-configure itself and can execute the DataSHIELD operations. Default is none.

env

The environment where DataSHIELD objects should be looked for: the DSLiteServer and the DSIConnection objects. Default is the Global environment.

Value

The login data for the datashield.login function.

See also

Examples

if (FALSE) { # \dontrun{
logindata <- setupSURVIVALTest()
conns <- datashield.login(logindata, assign=TRUE)
# do DataSHIELD analysis
datashield.logout(conns)
} # }